+++ groundguard/generate/__init__.py
@@ -1,7 +1,8 @@
 from groundguard.generate.grounded_generate import (
     GroundedResult,
     GroundingPolicyError,
+    grounded,
     grounded_generate,
 )
 
-__all__ = ["GroundedResult", "GroundingPolicyError", "grounded_generate"]
+__all__ = ["GroundedResult", "GroundingPolicyError", "grounded", "grounded_generate"]